fix: address review findings — stale cache, WS resilience, shared utils
